Done 4.21 and fixed postBlog without token errors

+++ backend/controllers/blog.js
@@ -9,8 +9,8 @@ blogRouter.get('/', async (request, response) => {
 })
 
 blogRouter.post('/', async (request, response, next) => {
+  if (!request.token) return response.status(401).json({ error: 'token missing' })
   const body = request.body
-  console.log(request)
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
@@ -36,8 +36,20 @@ blogRouter.post('/', async (request, response, next) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  if (!request.token) return response.status(401).json({ error: 'token missing' })
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blogpost = await Blog.findById(request.params.id)
+
+  if ( blogpost.user.toString() === user.id.toString() ){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(400)
+  }
 })
 
 blogRouter.put('/:id', async (request, response, next) => {
